Refactor CombatLog elseifs into polymorphic CombatEvent::CombatLogDescription
Create a virtual CombatEvent::CombatLogDescription.

Improves code localization as Description of each event is in the same
file as its constructor.

Simplify adding new events, and nested detail descriptions.
